change the lm dataset dir
