+++ paddlespeech/s2t/io/collator.py
@@ -19,7 +19,6 @@ from yacs.config import CfgNode
 
 from paddlespeech.s2t.frontend.augmentor.augmentation import AugmentationPipeline
 from paddlespeech.s2t.frontend.featurizer.speech_featurizer import SpeechFeaturizer
-from paddlespeech.s2t.frontend.featurizer.text_featurizer import TextFeaturizer
 from paddlespeech.s2t.frontend.normalizer import FeatureNormalizer
 from paddlespeech.s2t.frontend.speech import SpeechSegment
 from paddlespeech.s2t.frontend.utility import IGNORE_ID
@@ -44,43 +43,6 @@ def _tokenids(text, keep_transcription_text):
 
     tokens = np.array(tokens, dtype=np.int64)
     return tokens
-
-
-class TextCollatorSpm():
-    def __init__(self, unit_type, vocab_filepath, spm_model_prefix):
-        assert (vocab_filepath is not None)
-        self.text_featurizer = TextFeaturizer(
-            unit_type=unit_type,
-            vocab_filepath=vocab_filepath,
-            spm_model_prefix=spm_model_prefix)
-        self.eos_id = self.text_featurizer.eos_id
-        self.blank_id = self.text_featurizer.blank_id
-
-    def __call__(self, batch):
-        """
-        return type  [List, np.array [B, T], np.array [B, T], np.array[B]]
-        """
-        keys = []
-        texts = []
-        texts_input = []
-        texts_output = []
-        text_lens = []
-
-        for idx, item in enumerate(batch):
-            key = item.split(" ")[0].strip()
-            text = " ".join(item.split(" ")[1:])
-            keys.append(key)
-            token_ids = self.text_featurizer.featurize(text)
-            texts_input.append(
-                np.array([self.eos_id] + token_ids).astype(np.int64))
-            texts_output.append(
-                np.array(token_ids + [self.eos_id]).astype(np.int64))
-            text_lens.append(len(token_ids) + 1)
-
-        ys_input_pad = pad_list(texts_input, self.blank_id).astype(np.int64)
-        ys_output_pad = pad_list(texts_output, self.blank_id).astype(np.int64)
-        y_lens = np.array(text_lens).astype(np.int64)
-        return keys, ys_input_pad, ys_output_pad, y_lens
 
 
 class SpeechCollatorBase():
